% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castor_assign_variable_labels.R
\name{castor_assign_variable_labels}
\alias{castor_assign_variable_labels}
\title{Assign variable labels and variable classes}
\usage{
castor_assign_variable_labels(dat = dat)
}
\arguments{
\item{dat}{}
}
\description{
Assign variable labels and variable classes
}
