% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castor_collect_factor_groups.R
\name{castor_collect_factor_groups}
\alias{castor_collect_factor_groups}
\title{Collect the Castor factor groups for variable sets}
\usage{
castor_collect_factor_groups(dat = dat)
}
\arguments{
\item{dat}{}
}
\description{
Collect the Castor factor groups for variable sets
}
