% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castor_format_R.R
\name{castor_format_R}
\alias{castor_format_R}
\title{Read and format Castor CRF export data.}
\usage{
castor_format_R(fn = "CRF_export.xlsx")
}
\arguments{
\item{fn}{CRF_export.xlsx or other Castor project MS Excel filename}
}
\description{
Read and format Castor CRF export data.
}
