% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castor_read_data.R
\name{castor_read_data}
\alias{castor_read_data}
\title{Read the Castor CRF exported file for formatting}
\usage{
castor_read_data(fn = "CRF_export.xlsx")
}
\arguments{
\item{fn}{CRF_export.xlsx or other Castor project MS Excel filename}
}
\description{
Read the Castor CRF exported file for formatting
}
